Moved run_system command / test over to system
