+++ crates/bevy_app/src/system_registry.rs
@@ -1,6 +1,6 @@
 use crate::App;
 use bevy_ecs::prelude::*;
-use bevy_ecs::system::{SystemId, SystemRegistryError};
+use bevy_ecs::system::{RunSystem, SystemId, SystemRegistryError};
 
 impl App {
     /// Register a system with any number of [`SystemLabel`]s.
@@ -16,7 +16,7 @@ impl App {
 
     /// Runs the supplied system on the [`World`] a single time.
     ///
-    /// Calls [`SystemRegistry::run_system`](bevy_ecs::system::SystemRegistry::run_system).
+    /// Calls [`RunSystem::run_system`](bevy_ecs::system::RunSystem::run_system).
     #[inline]
     pub fn run_system<M, S: IntoSystem<(), (), M> + 'static>(&mut self, system: S) -> &mut Self {
         self.world.run_system(system);
